Currently it takes 4.27s to process the valid_map.tx:
```
$ ruby lib/continents_finder_runner.rb test/input/valid-map.txt  
4.27s user 0.10s system 97% cpu 4.458 total
``

We can optimise ```add_to_continent``` so it doesn't loop through the lands array all the time.
``` 
$ ruby lib/continents_finder_runner.rb test/input/valid-map.txt  
2.19s user 0.06s system 97% cpu 2.311 total
```